btrfs: move to fn pointers for get/put workspaces
The last patch created generic helpers for get_workspace() and
put_workspace(). This moves object ownership of the workspace_manager to
be in the specific compression file as compression code itself doesn't
care who its getting a workspace from.

This patch migrates ownership of the workspace_manager to the
compression type code rather than in the generic compression interface.
The init/cleanup and get/put methods are abstracted so each compression
algorithm can decide how they want to manage the workspaces.

Signed-off-by: Dennis Zhou <dennis@kernel.org>