btrfs: add zstd compression level support
Zstd compression requires different amounts of memory for each level of
compression. The prior patches implemented indirection to allow for each
compression type to manage their workspaces independently.

V3 moves away from the using set_level() to resize workspaces in favor
of just allocating a workspace of the appropriate level and using a
timer to reclaim unused workspaces. Compression is handled by a
workqueue which creates an implicit limit of the max number of
workspaces per level. The max level workspace is still allocated to
provide a route for forward progress. It is hidden from the LRU to
prevent reclamation of it.

When getting a workspace, it uses a bitmap to identify the levels that
are populated and scans up. If it finds a workspace that is greater than
it, it uses it but does not update the last_used time and the
corresponding place in the LRU. This provides a mechanism to decrease
memory utilization as we only keep around workspaces that are sized
appropriately for the in use compression levels.

By knowing which compression levels have available workspaces, we can
recycle rather than always create new workspaces as well as take
advantage of the preallocated max level for forward progress. If we hit
memory pressure, we sleep on the max level workspace. We continue to
rescan in case we can use a smaller workspace, but eventually should be
able to obtain the max level workspace or allocate one again should
memory pressure subside.

Taken from v2 [2]:
  Benchmarks run in qemu on Intel x86 with a single core.
  The benchmark measures the time to copy the Silesia corpus [0] to
  a btrfs filesystem 10 times, then read it back.

  The two important things to note are:
  - The decompression speed and memory remains constant.
  The memory required to decompress is the same as level 1.
  - The compression speed and ratio will vary based on the source.

  Level	Ratio	Compression	Decompression	Compression Memory
  1    	2.59 	153 MB/s   	112 MB/s     	0.8 MB
  2    	2.67 	136 MB/s   	113 MB/s     	1.0 MB
  3    	2.72 	106 MB/s   	115 MB/s     	1.3 MB
  4    	2.78 	86  MB/s   	109 MB/s     	0.9 MB
  5    	2.83 	69  MB/s   	109 MB/s     	1.4 MB
  6    	2.89 	53  MB/s   	110 MB/s     	1.5 MB
  7    	2.91 	40  MB/s   	112 MB/s     	1.4 MB
  8    	2.92 	34  MB/s   	110 MB/s     	1.8 MB
  9    	2.93 	27  MB/s   	109 MB/s     	1.8 MB
  10   	2.94 	22  MB/s   	109 MB/s     	1.8 MB
  11   	2.95 	17  MB/s   	114 MB/s     	1.8 MB
  12   	2.95 	13  MB/s   	113 MB/s     	1.8 MB
  13   	2.95 	10  MB/s   	111 MB/s     	2.3 MB
  14   	2.99 	7   MB/s   	110 MB/s     	2.6 MB
  15   	3.03 	6   MB/s   	110 MB/s     	2.6 MB

  [0] http://sun.aei.polsl.pl/~sdeor/index.php?page=silesia

[1] https://lore.kernel.org/linux-btrfs/20181031181108.289340-1-terrelln@fb.com/

Signed-off-by: Dennis Zhou <dennis@kernel.org>
Cc: Nick Terrell <terrelln@fb.com>
Cc: Omar Sandoval <osandov@fb.com>